+++ lab_1/task_2/decode.txt
@@ -1,0 +1,1 @@
+УДАЧНЫЕ АЛГОРИТМЫ СЖАТИЯ ИЛИ КАК ЕЩЕ ГОВОРЯТ КОМПРЕССИИ ПОЗВОЛЯЮТ НЕ ТОЛЪКО ЭФФЕКТИВНО ХРАНИТЪ КОНФИДЕНИАЛЪНЫЕ ДАННЫЕ НО И НАПРИМЕР ЗНАЧИТЕЛЪНО УМЕНЪШИТЪ РАЗМЕР ПРОГРАММЫ КОТОРУЮ ЗА ОДИН ПРИЕМ МОЖНО ЗАГРУЗИТЪ В ПАМЯТЪ И ИСПОЛНИТЪ СУЩЕСТВУЕТ НЕ МЕНЪШЕ ДЮЖИНЫ РАЗЛИЧНЫХ УПАКОВЩИКОВ ИСПОЛНИМЫХ ФАЙЛОВ НЕКОТОРЫЕ ИЗ НИХ СОДЕРЖАТ РЕАЛИЗАИ АЛГОРИТМОВ ШИФРОВАНИЯ С ЕЛЪЮ УСЛОЖНИТЪ ЖИЗНЪ ПОТЕНИАЛЪНЫМ ХАКЕРАМ НО НИ В ОДНОЙ ИЗ НИХ АЛГОРИТМ ШИФРОВАНИЯ НИКАК НЕ СВЯЗАН С АЛГОРИТМОМ СЖАТИЯ ОНИ РЕАЛИЗОВАНЫ КАК ОТДЕЛЪНЫЕ АЛГОРИТМЫ И ИСПОЛЪЗУЮТСЯ РАЗДЕЛЪНО
