+++ lab_1/task_2/key.json
@@ -1,0 +1,32 @@
+{
+    "Х":" ",
+    "Ч":"А",
+    "Щ":"В",
+    "Ъ":"Г",
+    "w":"Д",
+    "Ы":"Е",
+    "Ь":"Ж",
+    "Я":"З",
+    " ":"И",
+    "А":"Й",
+    "Б":"К",
+    "1":"Л",
+    "2":"М",
+    "Д":"Н",
+    "Е":"О",
+    "r":"П",
+    "t":"Р",
+    "И":"С",
+    "Й":"Т",
+    "К":"У",
+    "Л":"Ф",
+    "М":"Х",
+    "5":"Ч",
+    "О":"Ш",
+    "П":"Щ",
+    "7":"Ъ",
+    ">":"Ы",
+    "8":"Э",
+    "У":"Ю",
+    "Ф":"Я"
+}
